fix(security): move createOrgForUser out of "use server" module
Next.js auto-registers every export from a `"use server"` module as a
publicly invokable server action — reachable from any caller that
obtains the action ID, with no session check beyond what the function
performs itself. `createOrgForUser` lived in
`apps/web/app/(app)/complete-profile/actions.ts` (which begins with
`"use server";`) and took `userId` as a parameter without any auth
check. The function was therefore an unauthenticated endpoint that any
caller could use to create owner-organizations for ARBITRARY user IDs.

Concrete impact: an attacker scrapes / replays the action ID, then
calls `createOrgForUser(victimUserId, "spam")` repeatedly until the
victim hits `MAX_OWNED_ORGS_PER_USER` (= 3). The victim's account is
now stuck — they can't create their own organizations, and each
spam-org consumes the $150 free-credit grant.

Fix: move the function to a plain server lib at `apps/web/lib/org-create.ts`
(no `"use server"` directive → not a routable server action). The
three legitimate callers — `completeProfile` (the form action that
ran the function with `session.user.id`), `layout.tsx` (auto-create
on first sign-in, also session-authenticated), and
`/api/cli/auth/orgs` (route handler, session-checked) — all import
from the new location. Behavior at those call sites is unchanged;
they were already passing the session-derived userId.

The unauthenticated endpoint stops existing.
